Moved icon styles to the shared file
