+++ packages/react-components/react-breadcrumb-preview/src/components/BreadcrumbItem/useBreadcrumbItemStyles.styles.ts
@@ -2,18 +2,11 @@ import { makeStyles, mergeClasses, shorthands } from '@griffel/react';
 import type { BreadcrumbItemSlots, BreadcrumbItemState } from './BreadcrumbItem.types';
 import type { SlotClassNames } from '@fluentui/react-utilities';
 import { tokens, typographyStyles } from '@fluentui/react-theme';
+import { useIconStyles } from '../../utils/sharedStyles.styles';
 
 export const breadcrumbItemClassNames: SlotClassNames<BreadcrumbItemSlots> = {
   root: 'fui-BreadcrumbItem',
   icon: 'fui-BreadcrumbItem__icon',
-};
-
-/**
- * CSS variable names used internally for styling in the Breadcrumb.
- */
-export const breadcrumbCSSVars = {
-  breadcrumbIconSizeVar: '--fui-Breadcrumb--icon-size',
-  breadcrumbIconLineHeightVar: '--fui-Breadcrumb--icon-line-height',
 };
 
 /**
@@ -59,28 +52,6 @@ const useStyles = makeStyles({
   },
   noSpacing: {
     ...shorthands.padding(0),
-  },
-});
-
-const useIconStyles = makeStyles({
-  base: {
-    fontSize: `var(${breadcrumbCSSVars.breadcrumbIconSizeVar})`,
-    height: `var(${breadcrumbCSSVars.breadcrumbIconSizeVar})`,
-    lineHeight: `var(${breadcrumbCSSVars.breadcrumbIconLineHeightVar})`,
-    width: `var(${breadcrumbCSSVars.breadcrumbIconSizeVar})`,
-    marginRight: tokens.spacingHorizontalXS,
-  },
-  small: {
-    [breadcrumbCSSVars.breadcrumbIconSizeVar]: '12px',
-    [breadcrumbCSSVars.breadcrumbIconLineHeightVar]: tokens.lineHeightBase200,
-  },
-  medium: {
-    [breadcrumbCSSVars.breadcrumbIconSizeVar]: '16px',
-    [breadcrumbCSSVars.breadcrumbIconLineHeightVar]: tokens.lineHeightBase400,
-  },
-  large: {
-    [breadcrumbCSSVars.breadcrumbIconSizeVar]: '20px',
-    [breadcrumbCSSVars.breadcrumbIconLineHeightVar]: tokens.lineHeightBase600,
   },
 });
